+++ pso-flint-lucene/src/test/java/org/pageseeder/flint/lucene/facet/StringIntervalFacetTest.java
@@ -41,7 +41,7 @@ public class StringIntervalFacetTest {
   @BeforeClass
   public static void init() {
     // clean up previous test's data
-    for (File f : indexRoot.listFiles()) f.delete();
+    if (indexRoot.listFiles() != null) for (File f : indexRoot.listFiles()) f.delete();
     indexRoot.delete();
     try {
       index = new LuceneLocalIndex(indexRoot, "stringinterval", new StandardAnalyzer(), documents);
