+++ pso-flint-lucene/src/test/java/org/pageseeder/flint/lucene/query/SuggestionQueryTest.java
@@ -37,7 +37,7 @@ public class SuggestionQueryTest {
   @BeforeClass
   public static void init() {
     // clean up previous test's data
-    for (File f : indexRoot.listFiles()) f.delete();
+    if (indexRoot.listFiles() != null) for (File f : indexRoot.listFiles()) f.delete();
     indexRoot.delete();
     try {
       index = new LuceneLocalIndex(indexRoot, "suggestion", new StandardAnalyzer(), documents);
